requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
loguru~=0.6.0
celery~=5.2.7
redis==4.5.4
cloudscraper~=1.2.69
lxml~=4.9.2
aiohttp~=3.8.4
asyncio~=3.4.3
bs4~=0.0.1
beautifulsoup4~=4.12.0
django-filter~=23.1
